[wip][fix] fix judge pointer

+++ include/cc_sakura.h
@@ -111,7 +111,7 @@ struct Type{
 	TypeKind ty;
 	Type	 *ptr_to;
 	Member   *member;
-	size_t   index_size;
+	int      index_size;
 };
 
 // tree object
